+++ Source/AllowanceApp.Core/Services/AccountService.cs
@@ -21,13 +21,13 @@ namespace AllowanceApp.Core.Services
         public async Task<AllowancePoint> IncOrDecPointAsync(int id, string category, bool incrementing) =>
             await _actor.IncOrDecPointAsync(id, category, incrementing);
 
-        public async Task<AllowancePoint> UpdateAllowancePriceAsync(int id, string category, double amount) =>
+        public async Task<AllowancePoint> UpdateAllowancePriceAsync(int id, string category, int amount) =>
             await _actor.UpdateAllowancePriceAsync(id, category, amount);
 
         public async Task<Account> PayAllowanceAsync(int id) =>
             await _actor.PayAllowanceAsync(id);
 
-        public async Task<Account> ApplyTransactionAsync(int id, double amount, bool isWithdrawal, string? description) =>
+        public async Task<Account> ApplyTransactionAsync(int id, int amount, bool isWithdrawal, string? description) =>
             await _actor.ApplyTransactionAsync(id, amount, isWithdrawal, description);
 
         public async Task<string> DeleteAccountAsync(int id) =>
